Make FLINT_ABS safe by outputting ulong

+++ flint.h
@@ -293,7 +293,7 @@ typedef __mpfr_struct flint_mpfr;
 
 #define FLINT_MAX(x, y) ((x) > (y) ? (x) : (y))
 #define FLINT_MIN(x, y) ((x) > (y) ? (y) : (x))
-#define FLINT_ABS(x) ((slong)(x) < 0 ? (-(x)) : (x))
+#define FLINT_ABS(x) ((slong)(x) < 0 ? -((ulong) (x)) : (ulong) (x))
 #define FLINT_SIGN_EXT(x) (-(ulong)((slong)(x) < 0))
 #define FLINT_SGN(x) ((0 < (slong)(x)) - ((slong)(x) < 0))
